requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
deprecation==2.1.0
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
Jinja2==3.1.4
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-keycloak==4.4.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.37.2
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.43.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
